Refactor async_update to use asyncio
The `async_update` method in the solis_modbus sensor has been refactored to utilize asyncio for entity updates. Instead of using simple loops to call the update method on each entity, asyncio.gather and asyncio.to_thread are used to make the update calls asynchronously. This change should improve the efficiency of the update method, reducing the time taken and resources used.